Update DatabaseHandeller.java
Badr's Functions need to check on them, they have a problem in the Sub tasks and task how to differentiate between them by id?
and why the functions need to check on? because the connection doesn't work on laptop and i don't know why?. so, we can test them 
at the end. 
Regards. 
Fuck to the Fuckin' Society  ^_^